# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: El Príncipe Mestizo - Citizen journalism platform API
servers:
  - url: /api/v1
    description: Canonical API v1 path
  - url: /api
    description: Legacy (deprecated — será removido en v2)
...
paths:
...
  /admin/comments:
    get:
      operationId: adminGetComments
      tags: [admin]
      summary: Get all comments with filter
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, approved, pending]
      responses:
        "200":
          description: Comments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/comments/{id}/approve:
    patch:
      operationId: adminApproveComment
      tags: [admin]
      summary: Approve a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/comments/{id}:
    delete:
      operationId: adminDeleteComment
      tags: [admin]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        articleId:
          type: integer
        articleTitle:
          type: ["string", "null"]
        authorName:
          type: string
        authorEmail:
          type: ["string", "null"]
        content:
          type: string
        approved:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - articleId
        - authorName
        - content
        - approved
        - createdAt
...